# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API документация для e-commerce",
        "description": "Эта документация описывает REST API для интернет-магазина",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Локальный сервер"
        }
    ],
    "paths": {
...
        "/api/cart/checkout": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Оформить заказ (оплатить корзину)",
                "operationId": "9b93691975678b12acba054a7bc9aa9e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "payment_method_id"
                                ],
                                "properties": {
                                    "payment_method_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Заказ успешно создан"
                    },
                    "400": {
                        "description": "Корзина пуста"
                    },
                    "500": {
                        "description": "Ошибка при создании заказа"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{orderId}/status": {
            "patch": {
                "tags": [
                    "Orders"
                ],
                "summary": "Обновить статус заказа (Оплачен)",
                "operationId": "a8ee4b0b3958767efa744a601ce5e05c",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "ID заказа",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Статус заказа обновлен"
                    },
                    "400": {
                        "description": "Статус заказа нельзя изменить"
                    },
                    "404": {
                        "description": "Заказ не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Получить список заказов пользователя",
                "operationId": "eb9c6f17933fc485d7521de429b90db9",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Фильтрация по статусу заказа",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "На оплату",
                                "Оплачен",
                                "Отменен"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список заказов"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/orders/{orderId}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Получить заказ по ID",
                "operationId": "3ac61fe71b6874c8679de385a2d3da61",
                "parameters": [
                    {
                        "name": "orderId",
                        "in": "path",
                        "description": "ID заказа",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о заказе"
                    },
                    "404": {
                        "description": "Заказ не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}